requirements.txt
Babel==2.14.0
pandas==2.0.3
Pillow==10.1.0
Pillow==10.2.0
plotly==5.19.0
seaborn==0.13.2
streamlit==1.31.1
